WD-58 Moved passport stuff out of `server.js`
