+++ src/server/server.js
@@ -4,15 +4,14 @@ import webpack from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
 import express from 'express';
-import webpackConfig from '../../webpack.config.dev';
-import ajaxProxyRouter from './lib/ajaxProxyRouter';
 import morgan from 'morgan';
 import cookieParser from 'cookie-parser';
 import session from 'express-session';
 import passport from 'passport';
-import { OAuth2Strategy as GoogleStrategy } from 'passport-google-oauth';
-import authConfig from './auth/credentials';
-import { upsert } from './lib/util';
+
+import webpackConfig from '../../webpack.config.dev';
+import ajaxProxyRouter from './lib/ajaxProxyRouter';
+import passportGoogle from './auth/strategies/google';
 
 delete process.env.BROWSER;
 
@@ -21,45 +20,6 @@ const port = isDeveloping ? 3000 : process.env.PORT;
 const server = express();
 
 // TODO: write tests for server.js
-
-/**
- ******************************************
- passport stuff - TODO: Move into separate file
- ******************************************
- */
-passport.serializeUser((user, done) => {
-  const sessionUser = {
-    first_name: user.first_name,
-    last_name: user.last_name,
-    email: user.email,
-    photo_url: user.photo_url,
-  };
-  done(null, sessionUser);
-});
-
-passport.deserializeUser((sessionUser, done) => done(null, sessionUser));
-
-passport.use(new GoogleStrategy(
-  authConfig.google,
-  (accessToken, refreshToken, profile, done) => {
-    const params = {
-      email: profile.emails[0].value,
-      external_auth_type: 'google',
-    };
-    const data = {
-      first_name: profile.name.givenName,
-      last_name: profile.name.familyName,
-      email: profile.emails.length && profile.emails[0].value,
-      photo_url: profile.photos.length && profile.photos[0].value,
-      external_auth_type: 'google',
-      external_auth_id: profile.id,
-    };
-
-    upsert('/users', params, data)
-      .then(resp => done(null, resp))
-      .catch(err => done(err));
-  }
-));
 
 /**
  ******************************************
@@ -97,7 +57,6 @@ server.use(session({
 }));
 
 server.use(passport.initialize());
-
 server.use(passport.session());
 
 /**
@@ -105,14 +64,12 @@ server.use(passport.session());
  * user auth endpoints
  ******************************************
  */
-
-
 server.get('/auth/google',
-  passport.authenticate('google', { scope: ['openid email profile'] })
+  passportGoogle.authenticate('google', { scope: ['openid email profile'] })
 );
 
 server.get('/auth/google/callback',
-  passport.authenticate('google', { failureRedirect: '/' }),
+  passportGoogle.authenticate('google', { failureRedirect: '/' }),
   (req, res) => {
     res.redirect('/');
   }
